+++ src/zhinst/toolkit/driver/modules/__init__.py
@@ -9,6 +9,9 @@ from zhinst.toolkit.driver.modules.scope_module import ScopeModule
 from zhinst.toolkit.driver.modules.impedance_module import ImpedanceModule
 from zhinst.toolkit.driver.modules.device_settings_module import DeviceSettingsModule
 from zhinst.toolkit.driver.modules.pid_advisor_module import PIDAdvisorModule
+from zhinst.toolkit.driver.modules.precompensation_advisor_module import (
+    PrecompensationAdvisorModule,
+)
 
 ModuleType = t.Union[
     BaseModule,
@@ -19,6 +22,7 @@ ModuleType = t.Union[
     ImpedanceModule,
     DeviceSettingsModule,
     PIDAdvisorModule,
+    PrecompensationAdvisorModule,
 ]
 
 __all__ = [
@@ -31,4 +35,5 @@ __all__ = [
     "ImpedanceModule",
     "DeviceSettingsModule",
     "PIDAdvisorModule",
+    "PrecompensationAdvisorModule",
 ]