+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 0.5.1
 * Added full support for the following LabOne modules (no need to fallback to zhinst.core):
   * Impedance Module
+  * Precompensation Advisor Module
 
 ## Version 0.5.0
 * Added full support for the following LabOne modules (no need to fallback to zhinst.core):